requirements.txt
pymongo==4.7.3
python-dotenv==1.0.1
autopep8==2.0.1
pep8==1.7.1
flake8==7.0.0
pytest==8.0.0
pytest-asyncio==0.23.7
aiogram==2.9
